requirements.txt
numpy>=1.20.2
pandas>=1.2.4
matplotlib>=3.4.1
scikit-learn>=0.24.2
xgboost>=1.4.1
pyyaml>=5.4.1
seaborn>=0.11.1